Add files via upload
Folder with example files to use the code "PPI_node_size_by_centrality_and_color_assigned_by_OG". You just have to download this folder and change the directory indicated in the code to the directory where you downloaded the folder. This example contains the data for the construction of the ppi networks that correspond to three different species.